Add Butterworth & Gaussian

+++ Ass1/Distance.cpp
@@ -6,7 +6,7 @@ using namespace cv;
 
 
 
-static int D(int i, int j, Point p)
+static double D(int i, int j, Point p)
 {
 	return sqrt(pow(i-p.y,2)+pow(j-p.x,2));
 }
